extracted LP methods to utils
